Add new annotation move tests to where they belong
