+++ src/test/java/com/fasterxml/jackson/databind/tofix/CreatorFallback2438Test.java
@@ -7,6 +7,7 @@ import com.fasterxml.jackson.annotation.JsonProperty;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.databind.testutil.DatabindTestUtil;
+import com.fasterxml.jackson.databind.testutil.failing.JacksonTestFailureExpected;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
@@ -33,6 +34,7 @@ class CreatorFallback2438Test extends DatabindTestUtil {
 
     private final ObjectMapper MAPPER = newJsonMapper();
 
+    @JacksonTestFailureExpected
     @Test
     void creator2438() throws Exception {
         // note: by default, duplicate-detection not enabled, so should not
