+++ src/test/java/com/fasterxml/jackson/databind/tofix/JacksonInject2465Test.java
@@ -6,6 +6,7 @@ import com.fasterxml.jackson.annotation.*;
 import com.fasterxml.jackson.databind.InjectableValues;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.databind.testutil.DatabindTestUtil;
+import com.fasterxml.jackson.databind.testutil.failing.JacksonTestFailureExpected;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertNotNull;
@@ -45,6 +46,7 @@ class JacksonInject2465Test extends DatabindTestUtil {
     }
 
     // [databind#2465]
+    @JacksonTestFailureExpected
     @Test
     void injectWithCreator() throws Exception {
         ObjectMapper mapper = jsonMapperBuilder()
