+++ src/test/java/com/fasterxml/jackson/databind/tofix/JacksonInject4218Test.java
@@ -8,6 +8,7 @@ import com.fasterxml.jackson.annotation.JsonProperty;
 
 import com.fasterxml.jackson.databind.*;
 import com.fasterxml.jackson.databind.testutil.DatabindTestUtil;
+import com.fasterxml.jackson.databind.testutil.failing.JacksonTestFailureExpected;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
@@ -49,6 +50,7 @@ class JacksonInject4218Test extends DatabindTestUtil
     }
 
     // [databind#4218]
+    @JacksonTestFailureExpected
     @Test
     void injectFail4218() throws Exception
     {
