+++ src/test/java/com/fasterxml/jackson/databind/tofix/MapEntryFormat1419Test.java
@@ -8,6 +8,7 @@ import org.junit.jupiter.api.Test;
 import com.fasterxml.jackson.annotation.JsonFormat;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.databind.testutil.DatabindTestUtil;
+import com.fasterxml.jackson.databind.testutil.failing.JacksonTestFailureExpected;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
@@ -29,6 +30,7 @@ class MapEntryFormat1419Test extends DatabindTestUtil {
 
     private final ObjectMapper MAPPER = new ObjectMapper();
 
+    @JacksonTestFailureExpected
     @Test
     void wrappedAsObjectRoundtrip() throws Exception {
         BeanWithMapEntryAsObject input = new BeanWithMapEntryAsObject("foo", "bar");
