+++ src/test/java/com/fasterxml/jackson/databind/tofix/MapInclusion1649Test.java
@@ -9,6 +9,7 @@ import org.junit.jupiter.api.Test;
 import com.fasterxml.jackson.annotation.JsonInclude;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.databind.testutil.DatabindTestUtil;
+import com.fasterxml.jackson.databind.testutil.failing.JacksonTestFailureExpected;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
@@ -26,6 +27,7 @@ class MapInclusion1649Test extends DatabindTestUtil {
     final private ObjectMapper MAPPER = objectMapper();
 
     // [databind#1649]
+    @JacksonTestFailureExpected
     @Test
     void nonEmptyViaClass() throws IOException {
         // non-empty/null, include
