+++ src/test/java/com/fasterxml/jackson/databind/tofix/ObjectIdWithUnwrapping1298Test.java
@@ -13,6 +13,7 @@ import com.fasterxml.jackson.core.StreamWriteFeature;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.databind.json.JsonMapper;
 import com.fasterxml.jackson.databind.testutil.DatabindTestUtil;
+import com.fasterxml.jackson.databind.testutil.failing.JacksonTestFailureExpected;
 
 import static org.junit.jupiter.api.Assertions.assertNotNull;
 import static org.junit.jupiter.api.Assertions.fail;
@@ -53,6 +54,7 @@ class ObjectIdWithUnwrapping1298Test extends DatabindTestUtil {
         }
     }
 
+    @JacksonTestFailureExpected
     @Test
     void objectIdWithRepeatedChild() throws Exception {
         ObjectMapper mapper = JsonMapper.builder()
